requirements.txt
absl-py==0.11.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.0
bleach==3.1.4
cachetools==4.2.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-debug-toolbar==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.3.3
glob2==0.7
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-speech==2.0.1
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
libcst==0.3.17
librosa==0.6.3
llvmlite==0.32.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numba==0.49.0
numpy==1.16.0
olefile==0.46
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
proto-plus==1.13.0
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.8.6
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pydot==1.4.1
pydub==0.24.1
Pygments==2.6.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
requests==2.23.0
resampy==0.2.2
rsa==4.7.1
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==4.1.4
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
